删除很多不必要的类，简单整理了一下。

+++ core/src/com/mygdx/game/map/Mountain.java
@@ -12,7 +12,7 @@ public class Mountain{
 	 * @param num
 	 * @return */
 	float Displace(float num) {
-		float max=(float)num/(width+height)*3f;
+		float max= num /(width+height)*3f;
 		return (float)(Math.random()-0.5f)*max;
 	}
 	/**
